sep: http_block.cpp
